requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2

pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity